feature: #47 トップページだとリダイレクトで0の状態に切り戻していてあまり意味がないので切り戻す
